LayoutTests/fast/dom/dom-window-property-gc-after-frame-detach-expected.txt

Tests that Window properties do not get GC'd before their Window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frameWindow.applicationCache.foo is 1
PASS frameWindow.history.foo is 1
PASS frameWindow.location.foo is 1
PASS frameWindow.locationbar.foo is 1
PASS frameWindow.menubar.foo is 1
PASS frameWindow.navigator.foo is 1
PASS frameWindow.personalbar.foo is 1
PASS frameWindow.screen.foo is 1
PASS frameWindow.scrollbars.foo is 1
PASS frameWindow.statusbar.foo is 1
PASS frameWindow.toolbar.foo is 1

PASS frameWindow.applicationCache.foo is 1
PASS frameWindow.history.foo is 1
PASS frameWindow.location.foo is 1
PASS frameWindow.locationbar.foo is 1
PASS frameWindow.menubar.foo is 1
PASS frameWindow.navigator.foo is 1
PASS frameWindow.personalbar.foo is 1
PASS frameWindow.screen.foo is 1
PASS frameWindow.scrollbars.foo is 1
PASS frameWindow.statusbar.foo is 1
PASS frameWindow.toolbar.foo is 1

PASS frameWindow.applicationCache.foo is 1
PASS frameWindow.history.foo is 1
PASS frameWindow.location.foo is 1
PASS frameWindow.locationbar.foo is 1
PASS frameWindow.menubar.foo is 1
PASS frameWindow.navigator.foo is 1
PASS frameWindow.personalbar.foo is 1
PASS frameWindow.screen.foo is 1
PASS frameWindow.scrollbars.foo is 1
PASS frameWindow.statusbar.foo is 1
PASS frameWindow.toolbar.foo is 1
PASS successfullyParsed is true

TEST COMPLETE
